Move TestPreference.kt to top package
